requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
yfinance>=0.2.30
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.10
python-dotenv>=1.0.0
upstox-python-sdk>=2.0.0
websocket-client>=1.6.0
growwapi
